feat: Implement Quiz API with question generation and answering functionality
- Added quiz_api.dart to start the server and handle API requests.
- Removed obsolete QuestionAnswerDto and CustomException classes.
- Introduced new DTOs for question and error handling.
- Created QuizController to manage quiz-related operations.
- Developed QuizService for question generation and answer validation.
- Implemented user management with UserService and mock data.
- Established routes for quiz operations using Shelf Router.
- Updated pubspec.yaml and build.yaml for dependencies and configurations.
