+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.1.0
+
+- Updated dependencies to their latest versions.
+- Added `json_annotation`, `build_runner`, and `json_serializable` for JSON serialization support.
+- Refactored `CustomException` to be a concrete class instead of an abstract class.
+
+
 ## 1.0.0
 
 - Initial version.
