+++ analysis_options.yaml
@@ -13,6 +13,12 @@
 
 include: package:lints/recommended.yaml
 
+formatter:
+  page_width: 100
+
+
+analyzer:
+
 # Uncomment the following section to specify additional rules.
 
 # linter:
