+++ bin/quiz_api.dart
@@ -1,0 +1,6 @@
+import 'package:quiz_api/server.dart';
+
+void main(List<String> args) {
+  // Liga o servidor e inicia a API, o servidor ficará escutando as requisições e respondendo de acordo com as rotas implementadas.
+  startServer();
+}
