--- bin/src/exceptions/custom_exception.dart
@@ -1,9 +1,0 @@
-abstract class CustomException {
-  final int status;
-  final String message;
-
-  CustomException({
-    required this.status,
-    required this.message,
-  });
-}
