+++ lib/src/exceptions/custom_exception.dart
@@ -1,0 +1,7 @@
+/// 
+class CustomException implements Exception {
+  final int status;
+  final String message;
+
+  CustomException({required this.status, required this.message});
+}
