+++ lib/src/exceptions/not_found_exceptions.dart
