+++ lib/src/model/user_model.dart
