+++ pubspec.yaml
@@ -1,16 +1,19 @@
 name: quiz_api
 description: A server app using the shelf package and Docker.
-version: 1.0.0
+version: 1.1.0
 
 environment:
-  sdk: ^3.2.6
+  sdk: ^3.12.0
 
 dependencies:
-  args: ^2.4.0
-  shelf: ^1.4.0
-  shelf_router: ^1.1.0
+  args: ^2.7.0
+  shelf: ^1.4.2
+  shelf_router: ^1.1.4
+  json_annotation: ^4.12.0
 
 dev_dependencies:
-  http: ^1.1.0
-  lints: ^2.1.0
-  test: ^1.24.0
+  http: ^1.6.0
+  lints: ^6.1.0
+  test: ^1.31.1
+  build_runner: ^2.15.0
+  json_serializable: ^6.14.0
